requirements.txt
#librerias

django>=4.2,<5.0
djangorestframework
#Para que Django pueda leer .env (API Key Google Maps)
django-environ
#librerias para geolocalizacion convierte dirección en coordenadas
geopy
#libreria para hacer peticiones HTTP
requests
#para que Railway arranque la web
gunicorn
#para que funcione en Railway con BBDD PostgreSQL
psycopg2-binary
#libreria para manejar CORS (vercel lo necesita para que el frontend pueda hacer peticiones al backend)
django-cors-headers
#libreria para manejar autenticación con JWT
djangorestframework-simplejwt
#libreria para manejar URLs de bases de datos (railway permite Django para leer BBDD en PostgreSQL)
dj-database-url
#libreria para servir archivos estáticos en producción (railway)
whitenoise
